+++ contrib/libarchive/libarchive/archive_read_support_format_xar.c
@@ -2530,8 +2530,8 @@ xml_data(void *userData, const char *s, int len)
 #if DEBUG
 	{
 		char buff[1024];
-		if (len > sizeof(buff)-1)
-			len = sizeof(buff)-1;
+		if (len > (int)sizeof(buff)-1)
+			len = (int)sizeof(buff)-1;
 		memcpy(buff, s, len);
 		buff[len] = 0;
 		fprintf(stderr, "\tlen=%d:\"%s\"\n", len, buff);
